// graphql-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};
// ...
export type MarkdownRemark = Node & {
  id: Scalars['ID'];
  frontmatter?: Maybe<MarkdownRemarkFrontmatter>;
  excerpt?: Maybe<Scalars['String']>;
  rawMarkdownBody?: Maybe<Scalars['String']>;
  fileAbsolutePath?: Maybe<Scalars['String']>;
  fields?: Maybe<MarkdownRemarkFields>;
  html?: Maybe<Scalars['String']>;
  htmlAst?: Maybe<Scalars['JSON']>;
  excerptAst?: Maybe<Scalars['JSON']>;
  headings?: Maybe<Array<Maybe<MarkdownHeading>>>;
  timeToRead?: Maybe<Scalars['Int']>;
  tableOfContents?: Maybe<Scalars['String']>;
  wordCount?: Maybe<MarkdownWordCount>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type MarkdownRemarkFields = {
  slug?: Maybe<Scalars['String']>;
};
// ...
export type MarkdownRemarkFrontmatter = {
  title?: Maybe<Scalars['String']>;
  date?: Maybe<Scalars['Date']>;
  previewImage?: Maybe<File>;
  image?: Maybe<File>;
  preview?: Maybe<Scalars['String']>;
};
// ...
export type SitePageContext = {
  pageNumber?: Maybe<Scalars['Int']>;
  humanPageNumber?: Maybe<Scalars['Int']>;
  skip?: Maybe<Scalars['Int']>;
  limit?: Maybe<Scalars['Int']>;
  numberOfPages?: Maybe<Scalars['Int']>;
  previousPagePath?: Maybe<Scalars['String']>;
  nextPagePath?: Maybe<Scalars['String']>;
  slug?: Maybe<Scalars['String']>;
};
// ...
export type PostPreviewImageFragment = { fluid?: Maybe<GatsbyImageSharpFluid_NoBase64Fragment> };

export type PostImageFragment = { fluid?: Maybe<GatsbyImageSharpFluid_NoBase64Fragment> };

export type PostFragment = (
  Pick<MarkdownRemark, 'id'>
  & { content: MarkdownRemark['html'], preview: MarkdownRemark['excerpt'] }
  & { frontmatter?: Maybe<(
    Pick<MarkdownRemarkFrontmatter, 'title' | 'date'>
    & { previewImage?: Maybe<{ childImageSharp?: Maybe<PostPreviewImageFragment> }>, image?: Maybe<{ childImageSharp?: Maybe<PostImageFragment> }> }
  )>, fields?: Maybe<Pick<MarkdownRemarkFields, 'slug'>> }
);
// ...
export type PostsListPageQuery = { allMarkdownRemark: { posts: Array<{ post: PostFragment }> } };